Change index file
